+++ pioreactor/background_jobs/leader/mqtt_to_db_streaming.py
@@ -225,9 +225,11 @@ def parse_dosing_events(topic: str, payload: pt.MQTTMessagePayload) -> dict:
 
 
 def parse_experiment_profile_runs(topic: str, payload: pt.MQTTMessagePayload) -> dict:
+    metadata = produce_metadata(topic)
     return {
         "started_at": current_utc_datetime(),
         "experiment_profile_name": payload.decode("utf-8"),
+        "experiment": metadata.experiment,
     }